feat: log out center of xyz tile from cloudfront requests too
