+++ packages/shared/src/index.ts
@@ -8,4 +8,6 @@ export { LambdaHttpResponseAlb } from './lambda.response.alb';
 export { LambdaHttpResponseCloudFront, LambdaHttpResponseCloudFrontRequest } from './lambda.response.cf';
 export { LambdaHttpResponse, LambdaType } from './lambda.response.http';
 export { Logger } from './log';
+export { getXyzFromPath, PathData } from './path';
+export { Projection } from './projection';
 export { LambdaSession } from './session';
